// src/global/type.ts
export type AUTO_CENTER_JUSTIFY_LEFT_RIGHT =
  | 'auto'
  | 'center'
  | 'justify'
  | 'left'
  | 'right';
export type FLEX_JUSTIFYCONTENT =
  | 'center'
  | 'flex-end'
  | 'flex-start'
  | 'space-around'
  | 'space-between'
  | 'space-evenly';

export type FLEX_ALIGNITEMS =
  | 'center'
  | 'flex-end'
  | 'flex-start'
  | 'baseline'
  | 'stretch';

export type FLEXDIRECTION = 'column' | 'column-reverse' | 'row' | 'row-reverse';